pyproject.toml
[project]
name = "e-learning-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-forms>=1.1.8",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "rest-framework-simplejwt>=0.0.2",
    "ruff>=0.14.14",
]


[tool.ruff]

line-length = 88
target-version = "py311"


[tool.ruff.lint]

# Enable pyflakes (F), pycodestyle (E), isort (I)
select = ["E", "F", "I"]
ignore = ["F401", "E501"]  # Ignore unused imports everywhere



[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true



[tool.black]
max-line-length = 88 # Or your desired length
